+++ truncate-stream.js
@@ -8,8 +8,14 @@ function TruncateStream(opts) {
 
   this._truncateState = {
     bytes: 0,
-    maxBytes: opts.maxBytes
+    maxBytes: opts.maxBytes,
+    truncated: false,
+    src: []
   };
+
+  this.on('pipe', function (source) {
+    this._truncateState.src.push(source);
+  });
 }
 
 util.inherits(TruncateStream, stream.Transform);
@@ -17,11 +23,31 @@ module.exports = TruncateStream;
 
 TruncateStream.prototype._transform = function _transform(chunk, enc, callback) {
   var state = this._truncateState;
+  if (state.truncated) return callback();
+
   state.bytes += chunk.length;
 
   // are we there yet?
-  if (state.bytes < state.maxBytes) return this.push(chunk);
+  if (state.bytes < state.maxBytes) {
+    // nope
+    this.push(chunk);
+    return callback();
+  }
 
+  state.truncated = true;
+
+  // push exactly what's left to make up maxBytes
   this.push(chunk.slice(0, chunk.length - (state.bytes - state.maxBytes)));
-  this.push(null);
+
+  // unpipe ourselves and redirect their output to /dev/null
+  for (var i = 0; i < state.src.length; i++) {
+    state.src[i].unpipe(this);
+    state.src[i].on('data', devnull);
+  }
+  
+  state.src = null;
+  this.end();
+  callback();
 };
+
+function devnull() {}