+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # truncate-stream changelog
 
-## 1.0.0 (2014/11/12)
+## 1.0.1 (2014/11/12)
 
-  - initial version
+  - fix streams that emit more than 1 chunk
+
+## 1.0.0 (2014/11/12)
+*don't use; broken for streams that emit more than 1 chunk*